backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Image Processing
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
scikit-image==0.22.0

# AI/ML
mediapipe==0.10.9
tensorflow==2.15.0
torch==2.1.2
torchvision==0.16.2

# Face Detection and Analysis
face-recognition==1.3.0
dlib==19.24.2

# Cloud Storage
boto3==1.34.28
python-dotenv==1.0.0

# PDF Generation
reportlab==4.0.9
weasyprint==60.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
